refactor: made uutils.cpp with trivial functions
